Split counselor_home_screen.dart into components
