+++ pubspec.yaml
@@ -45,7 +45,7 @@ dependencies:
   http: ^1.4.0
   flutter_dotenv: ^5.2.1
   file_picker: ^10.1.3
-  intl: ^0.18.1
+  intl: 0.19.0
 
 dev_dependencies:
   flutter_test:
